add light color theme and CLI argument define it
