+++ src/ui/mod.rs
@@ -6,4 +6,4 @@ pub mod result_list;
 
 mod input_handler;
 mod scroll_offset_list;
-mod theme;
+pub mod theme;
